% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_prices.R
\name{get_current_spot}
\alias{get_current_spot}
\title{Get current spot prices}
\usage{
get_current_spot(types, region)
}
\arguments{
\item{types}{Vector of instance types}

\item{region}{AWS EC2 region}
}
\description{
Get the current spot prices for a collection of EC2 instance types in a given region.
}
\details{
Note that this will not get the current spot price for the instance type in
every availability zone in a region. Once it gets the spot price for at least
one availability zone, it will move on. A further enhancement would be to get spot
prices at all availability zones (or a specified subset of availability zones).
}
