% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_prices.R
\name{get_instance_resource_prices}
\alias{get_instance_resource_prices}
\title{Get instance prices}
\usage{
get_instance_resource_prices(memory, cpu, region, universe = NULL)
}
\arguments{
\item{memory}{Minimum memory (in GB) requested}

\item{cpu}{Minimum number of CPU cores requested}

\item{region}{AWS EC2 region}

\item{universe}{Universe of instances to consider. If null, will use memory and cpu to get the instance universe.}
}
\description{
Get the on-demand and current spot prices for a set of EC2 instances.
If the user provides memory (in GB) and CPU cores requested, it will get the
on-demand and spot prices for instances with at least that much memory and CPU.
If the user provides a universe of instances, it will get prices at that set of instances.
}
