% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_job.R
\name{get_instance_state}
\alias{get_instance_state}
\title{Get EC2 instance state}
\usage{
get_instance_state(ec2, instance_id)
}
\arguments{
\item{ec2}{An EC2 service objected created by \link[paws:ec2]{paws::ec2}}

\item{instance_id}{The id of the instance of which to get the status}
}
\value{
The instance state name
}
\description{
Gets the current state of an EC2 instance
}
