% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_prices.R
\name{get_instance_universe}
\alias{get_instance_universe}
\title{Get EC2 instance universe}
\usage{
get_instance_universe(memory, cpu, region)
}
\arguments{
\item{memory}{Minimum memory (in GB) requested}

\item{cpu}{Minimum number of CPU cores requested}

\item{region}{AWS EC2 region}
}
\description{
Get a data frame of EC2 instances with greater than or equal to a given
amount of memory and CPU in a given EC2 region.
}
\details{
Only instances types in the c5, m5, m5a, r5, r5a, t3, and t3a classes with less than
128GB of memory are included. For t3/t3a, only instances with less than or equal
to 4GB of memory are included. The 128GB memory cap is included to prevent inadvertently
choosing a massive instance.
}
