% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_prices.R
\name{get_optimal_ondemand}
\alias{get_optimal_ondemand}
\title{Get optimal on-demand instance}
\usage{
get_optimal_ondemand(memory, cpu, region, universe = NULL)
}
\arguments{
\item{memory}{Minimum memory (in GB) requested}

\item{cpu}{Minimum number of CPU cores requested}

\item{region}{AWS EC2 region}

\item{universe}{Universe of instances to consider. If null, will use memory and cpu to get the instance universe.}
}
\value{
A list with information about the lowest-priced instance satisfying the request
}
\description{
The user provides a desired minimum amount of memory and CPU, along with a region.
The function will find the lowest price on-demand instance having at least
that much memory and CPU.
}
