% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{list_r_job_instances}
\alias{list_r_job_instances}
\title{List running R jobs}
\usage{
list_r_job_instances(region, states = c("running", "pending"))
}
\arguments{
\item{region}{AWS region}

\item{states}{Instance states to include}
}
\value{
A data frame with information about running R jobs
}
\description{
List R jobs running on EC2 worker instances in a given region. It gets a list of
all running EC2 instances, applying a filter on the \code{JOB_TYPE} tag associated
with the EC2 instances. The \code{JOB_TYPE} tag is set to \verb{R worker job} when a job is
launched with \code{\link[=run_job]{run_job()}}.
}
