% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_job.R
\name{run_job}
\alias{run_job}
\title{Run R worker job}
\usage{
run_job(
  script,
  job_id,
  image_id,
  instance_type,
  spot,
  spot_max_price,
  efs_name,
  security_group_id,
  ssm_iam_profile,
  efs_mnt_pt,
  region,
  instance_max_hours,
  exec_timeout_hours,
  r_version,
  ssm_work_dir,
  log_file,
  rmd_opts = list(),
  status_sleep_seconds = 30
)
}
\arguments{
\item{script}{Path to the R script or R Markdown document to execute (on EFS storage)}

\item{job_id}{Job id}

\item{image_id}{id of the AMI to launch}

\item{instance_type}{Instance type}

\item{spot}{Boolean value for whether to launch a spot instance}

\item{spot_max_price}{If launching a spot instance, the maximum bid price}

\item{efs_name}{EFS id}

\item{security_group_id}{Security group id}

\item{ssm_iam_profile}{Name of IAM profile with SSM permissions. This profile will be
attached to the launched EC2 instance.}

\item{efs_mnt_pt}{Mount point of EFS storage on worker instance}

\item{region}{AWS EC2 region}

\item{instance_max_hours}{Maximum lifetime of the instance in hours before it will
automatically shut down}

\item{exec_timeout_hours}{Number of hours for timeout of the SSM command for executing
the script}

\item{r_version}{Version of R to use}

\item{ssm_work_dir}{Working directory of SSM command for executing R script/document}

\item{log_file}{Path of log file to which stdout/stderr of R script will be written}

\item{rmd_opts}{If running an R Markdown document, optional parameters for the output file path
and document parameters}

\item{status_sleep_seconds}{How many seconds to wait between checking the status of
the execution of the R script}
}
\description{
Spin up an AWS EC2 instance and run an R script or R Markdown document in that instance.
It launched an EC2 instance, attaches EFS storage, then send a command to execute the
R script or RMD file as an SSM command.
This function will block the R session until execution is complete.
}
