% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_job.R
\name{stop_instances}
\alias{stop_instances}
\title{Stop EC2 Instances}
\usage{
stop_instances(ids, region)
}
\arguments{
\item{ids}{Vector or list of ids of instances to stop}

\item{region}{Region of EC2 instances to stop}
}
\description{
Stop one or more running EC2 instances by instance id.
}
