% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_job.R
\name{stop_job}
\alias{stop_job}
\title{Stop an R job}
\usage{
stop_job(id)
}
\arguments{
\item{id}{The id of the job to stop}
}
\description{
\code{stop_job} stops a job that was started with \code{\link[=run_job_wrapper]{run_job_wrapper()}}.
The user must provide the id of the job to stop. This id is automatically
generated by \code{run_job_wrapper}. The id is printed in the output in the Jobs
pane in RStudio, and is (invisibly) returned by \code{run_job_wrapper}.
}
